- got the resize bug. turns out you couldn't resize the widgets due to a bug in the way the EwlFlag enumeration was declared.  stupid C -- fuck bitshifting, i'm using explicit hex now. (which i would have seen hours ago had i turned debugging on). - okay so  now the resize shit works, and hte resize shit in the button callback works, but only if you resize the window once. I'll tackle this one tomorrow.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@3173 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ misc/ewl/src/types.h
@@ -12,7 +12,7 @@ typedef unsigned int    EwlType;
 typedef EwlType         EwlWidgetType;
 typedef void            EwlData;
 typedef char            EwlBool;
-typedef unsigned int    EwlFlag;
+typedef unsigned long   EwlFlag;
 typedef unsigned int    EwlWidgetState;
 
 /* delta for doing double comparisons */
